+++ assets/controllers.js
@@ -1,5 +1,6 @@
 import { startStimulusApp } from '@symfony/stimulus-bundle';
 import BasicController from './js/basic.js';
+import EditShortcutController from './js/edit-shortcut.js';
 import LanguagesController from './js/languages.js';
 import MenuActiveController from './js/menu-active.js';
 
@@ -7,5 +8,6 @@ import MenuActiveController from './js/menu-active.js';
 globalThis.c975lStimulusApp ??= startStimulusApp();
 const app = globalThis.c975lStimulusApp;
 app.register('basic', BasicController);
+app.register('edit-shortcut', EditShortcutController);
 app.register('languages', LanguagesController);
 app.register('menu-active', MenuActiveController);